requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.0
aiosqlite==0.20.0
